pyproject.toml
[project]
name = "mi-mary"
version = "0.1.0"
description = ""
authors = [
    {name = "ALEXSTONE",email = "solo-drive@bk.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "mi"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
